% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\docType{class}
\name{Block-class}
\alias{Block-class}
\title{Block}
\description{
Class used within the package to keep the structure and information about
the generated blocks.
}
\section{Slots}{

\describe{
\item{\code{features}}{a vector of numeric which contains the indices of the block.}

\item{\code{explained_varaiance}}{a numeric which contains the variance explained of
the blocks variables based on the whole data set.}

\item{\code{is_valid}}{a logival which indicates if the block structure is valid.}
}}

