% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla-package.R
\docType{package}
\name{pla-package}
\alias{pla-package}
\alias{_PACKAGE}
\title{Principle Loading Analysis Package

This package is an implementation of the principle loading analysis.}
\description{
This package provides principle loading analysis, a method for reducing the number of features within a data set.
}
\references{
\insertRef{Bauer.19.02.2021}{pla}
\insertRef{Bauer.10.07.2020}{pla}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ronho/pla}
  \item Report bugs at \url{https://github.com/Ronho/pla/issues}
}

}
\author{
\strong{Maintainer}: Ron Holzapfel \email{ronholzapfel@outlook.de}

Authors:
\itemize{
  \item Jan Bauer [contributor]
}

}
\keyword{internal}
