% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.keep_blocks}
\alias{pla.keep_blocks}
\title{Keep Blocks

Used to remove each variable from the original data set which is part of
any of the blocks that have not been excluded by the block_indizes.}
\usage{
pla.keep_blocks(x, blocks, block_indizes)
}
\arguments{
\item{x}{data.frame matrix, the raw data; should
be the same used to obtain the blocks.}

\item{blocks}{list of blocks; obtained by the pla function.}

\item{block_indizes}{vector of numeric; indizes of the blocks from the
block list which variables should not be removed from the data set.}
}
\value{
data.frame, data without the removed variables
}
\description{
Keep Blocks

Used to remove each variable from the original data set which is part of
any of the blocks that have not been excluded by the block_indizes.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9),
d = c(10:12)
)
obj <- pla(data)
data <- pla.keep_blocks(data, obj$blocks, c(2, 3))
}
