% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.show}
\alias{pla.show}
\title{Show Function for pla S3}
\usage{
pla.show(x, ...)
}
\arguments{
\item{x}{a pla object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Shows the blocks, threshold, threshold_mode and the loadings.
}
\examples{
\dontrun{
require(AER)
data("OECDGrowth")

pla_obj = pla(OECDGrowth,cor=TRUE,thresholds = 0.5)
show(pla_obj)
}

}
