% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.spla}
\alias{pla.spla}
\title{Sparse Principal Loading Analysis}
\usage{
pla.spla(
  x,
  cor = FALSE,
  threshold = 0.01,
  threshold_mode = "cutoff",
  expvar = "approx",
  check = "rnc",
  orthogonal = TRUE,
  type = "data",
  lambda = 0,
  para = 0.5,
  ...
)
}
\arguments{
\item{x}{a numeric matrix or data frame which provides the data for the
sparse principal loading analysis.}

\item{cor}{a logical value indicating whether the calculation should use the
correlation or the covariance matrix. This option is only effective if the
parameter orthogonol is set to TRUE.}

\item{threshold}{a numeric value used to determine "small" values inside the
loadings. This option is only effective if the parameter orthogonol is set to FALSE.}

\item{threshold_mode}{a character string indicating how the threshold is
determined and used. \code{cutoff} indicates the usage of a threshold value.
\code{percentage} indicates that the cutoff value is determined by the maximum
element of each vector multiplied with the threshold value. This option is only
effective if the parameter orthogonol is set to FALSE.}

\item{expvar}{a character string indicating the method used for calculating
the explained variance. \code{approx} uses the explained variance of each
eigenvector i.e. its eigenvalue. \code{exact} uses the variance of each variable.}

\item{check}{a character string indicating if only rows or rows as well as columns
are used to detect the underlying block structure. \code{rows} checks if the rows fulfill
the required structure. \code{rnc} checks if rows and columns fulfill the required structure.}

\item{orthogonal}{a logical value indicating whether ...}

\item{type}{a character string indicating how ...}

\item{lambda}{a numeric value used to determine ...}

\item{para}{a numeric value used to determine ...}

\item{...}{further arguments passed to or from other methods.}
}
\value{
single or list of pla class containing the following attributes:
\item{x}{
  a numeric matrix or data frame which equals the input of \code{x}.
}
\item{loadings}{
  a matrix of variable loadings (i.e. a matrix containing the
  eigenvectors of the dispersion matrix).
}
\item{threshold}{
  a numeric value which equals the input of \code{thresholds}.
}
\item{threshold_mode}{
  a character string which equals the input of \code{threshold_mode}.
}
\item{blocks}{
  a list of blocks which are identified by principal loading analysis.
}
}
\description{
Sparse PLA
}
\examples{
if(requireNamespace("AER")){
require(AER)
}

}
