% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.thresholds}
\alias{pla.thresholds}
\title{Principle Loading Analysis - Thresholds

This function allows you to perform principle loading analysis on a data set
using multiple different thresholds. This allows you to determine the best 
threshold on a given data set.}
\usage{
pla.thresholds(x, thresholds, ...)
}
\arguments{
\item{x}{data.frame matrix, the raw data.}

\item{thresholds}{vector of numeric, determines "small" values inside
the eigen vectors.}

\item{...}{further arguments passed to pla.}
}
\value{
List of the following parameters which are provided as a pla class list:
\item{eigen_vectors}{matrix, eigen vector matrix obtained from \code{eigen}}
\item{threshold}{numeric, threshold}
\item{blocks}{list of blocks, blocks describing the explained variance}
}
\description{
Principle Loading Analysis - Thresholds

This function allows you to perform principle loading analysis on a data set
using multiple different thresholds. This allows you to determine the best 
threshold on a given data set.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9)
)
thresholds <- c(0.1, 0.2, 0.3, 0.4, 0.5)
pla.thresholds(data, thresholds = thresholds)

}
\seealso{
[pla()] for further information about pla.
}
