% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{print.pla}
\alias{print.pla}
\title{Print Function for pla S3}
\usage{
\method{print}{pla}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{\code{x}}{a pla object.}
}
\description{
Prints the blocks, threshold, threshold_mode and the loadings.
}
\examples{
require(AER)
data("OECDGrowth")

pla_obj = pla(OECDGrowth,cor=TRUE,thresholds = 0.5)
print(pla_obj)

}
