% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinvars-package.R
\docType{package}
\name{prinvars-package}
\alias{prinvars-package}
\alias{_PACKAGE}
\title{Principal Variables Package

This package provides methods for dimensionality reduction.}
\description{
Provides methods for reducing the number of features within a data set. See Bauer JO (2021) <doi:10.1145/3475827.3475832> and Bauer JO, Drabant B (2021) <doi:10.1016/j.jmva.2021.104754> for more information on principal loading analysis.
}
\references{
\insertRef{Bauer.06242021}{prinvars}
\insertRef{Bauer.2021}{prinvars}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ronho/prinvars}
  \item Report bugs at \url{https://github.com/Ronho/prinvars/issues}
}

}
\author{
\strong{Maintainer}: Ron Holzapfel \email{ronholzapfel@outlook.de}

Authors:
\itemize{
  \item Jan O. Bauer [contributor]
}

}
\keyword{internal}
