% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{show,Block-method}
\alias{show,Block-method}
\title{Block - Show

Prints the blocks structure.}
\usage{
\S4method{show}{Block}(object)
}
\arguments{
\item{object}{block.}
}
\description{
Block - Show

Prints the blocks structure.
}
\examples{
block <- new("Block", features = c(2, 5), explained_variance = 0.03)
print(block)
}
