% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{str,Block-method}
\alias{str,Block-method}
\title{Block - str

Generic function to create a string out of the blocks structure.}
\usage{
\S4method{str}{Block}(object)
}
\arguments{
\item{object}{block.}
}
\description{
Block - str

Generic function to create a string out of the blocks structure.
}
\examples{
block <- new("Block", indices = c(2, 5), explained_variance = 0.03)
str(block)
}
