% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\docType{class}
\name{Block-class}
\alias{Block-class}
\title{Block}
\description{
Classed used within the package to keep the structure and information about
the generated blocks. A Block is a a continuous sequence of 1s
within a vector.
}
\section{Slots}{

\describe{
\item{\code{rows}}{vector of numeric, indizes of the rows which belong to the block}

\item{\code{columns}}{vector of numeric, indizes of the columns which belong to the block}

\item{\code{explained_varaiance}}{numeric, explained variance of the blocks variables
based on the whole data set.}
}}

