% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\docType{class}
\name{Block-class}
\alias{Block-class}
\title{Block}
\description{
Classed used within the package to keep the structure and information about
the generated blocks. A Block is a a continuous sequence of 1s
within a vector.
}
\section{Slots}{

\describe{
\item{\code{start}}{numeric, beginning index of the sequence; -1 indicates that no 1 is found.}

\item{\code{end}}{numeric, ending index of the sequence; -1 indicates that no 1 is found.}

\item{\code{variables}}{vector of numeric, indizes of the variables having the same
start and end.}

\item{\code{explained_varaiance}}{numeric, explained variance of the blocks variables
based on the whole data set.}
}}

