% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\docType{class}
\name{Block-class}
\alias{Block-class}
\title{Block}
\description{
Classed used within the package to keep the structure and information about
the generated blocks.
}
\section{Slots}{

\describe{
\item{\code{indices}}{vector of numeric, indices which belong to the block}

\item{\code{explained_varaiance}}{numeric, explained variance of the blocks variables
based on the whole data set.}
}}

