% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.drop_blocks}
\alias{pla.drop_blocks}
\title{Drop Blocks

Used to remove each variable from the original data set which is part of any
of the blocks according to the passed indices.}
\usage{
pla.drop_blocks(object, block_indices, ...)
}
\arguments{
\item{object}{data.frame matrix, the raw data; should be the same used to
obtain the blocks.}

\item{block_indices}{list of numeric; indices of blocks that should be
dropped.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
list of the following attributes:
\item{x}{
  matrix, transformed matrix with removed attributes
}
\item{conditional_matrix}{
  matrix, conditional matrix
}
}
\description{
Drop Blocks

Used to remove each variable from the original data set which is part of any
of the blocks according to the passed indices.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9),
d = c(10:12)
)
obj <- pla(data)
data <- pla.drop_blocks(obj, c(1))

}
