% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.drop_blocks}
\alias{pla.drop_blocks}
\title{Drop Blocks

Used to remove each variable from the original data set which is part of
any of the passed blocks.}
\usage{
pla.drop_blocks(x, blocks)
}
\arguments{
\item{x}{data.frame matrix, the raw data; should
be the same used to obtain the blocks.}

\item{blocks}{list of blocks; blocks obtained by the pla function and should be dropped.}
}
\value{
data.frame, data without the removed variables
}
\description{
Drop Blocks

Used to remove each variable from the original data set which is part of
any of the passed blocks.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9),
d = c(10:12)
)
obj <- pla(data)
data <- pla.drop_blocks(data, obj$blocks[1])

}
