% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.keep_blocks}
\alias{pla.keep_blocks}
\title{Keep Blocks

Used to only keep each variable of the original data set which is part of
any of the passed blocks.}
\usage{
pla.keep_blocks(x, blocks)
}
\arguments{
\item{x}{data.frame matrix, the raw data; should be the same used to obtain the blocks.}

\item{blocks}{list of blocks; blocks that are obtained by the pla function and should
be kept.}
}
\value{
data.frame, data without the removed variables
}
\description{
Keep Blocks

Used to only keep each variable of the original data set which is part of
any of the passed blocks.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9),
d = c(10:12)
)
obj <- pla(data)
data <- pla.keep_blocks(data, obj$blocks[1])

}
