% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla.keep_blocks}
\alias{pla.keep_blocks}
\title{Keep Blocks

Used to only keep each variable of the original data set which is part of any
of the blocks according to the passed indices.}
\usage{
pla.keep_blocks(object, blocks, ...)
}
\arguments{
\item{object}{a pla object.}

\item{blocks}{a list of numeric values indicating the indices of the blocks
that should be kept.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
list of the following attributes:
\item{x}{
  a numeric matrix or data frame which equals the input data for the pla
  object without any feature that is not part of the blocks that should be
  kept.
}
\item{cc_matrix}{
  a numeric matrix or data frame which contains either the conditional
  covariance or correlation matrix.
}
}
\description{
Keep Blocks

Used to only keep each variable of the original data set which is part of any
of the blocks according to the passed indices.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9),
d = c(10:12)
)
obj <- pla(data)
data <- pla.keep_blocks(obj, c(1))

}
