% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla}
\alias{pla}
\title{Principal Loading Analysis}
\usage{
pla(
  x,
  cor = FALSE,
  scaled_ev = FALSE,
  thresholds = 0.33,
  threshold_mode = "cutoff",
  expvar = "approx",
  check = "rnc",
  ...
)
}
\arguments{
\item{x}{a numeric matrix or data frame which provides the data for the
principal loading analysis.}

\item{cor}{a logical value indicating whether the calculation should use the
correlation or the covariance matrix.}

\item{scaled_ev}{a logical value indicating whether the eigenvectors should
be scaled.}

\item{thresholds}{a numeric value or list of numeric values used to determine
"small" values inside the eigenvectors. If multiple values are given, a list
of pla results will be returned.}

\item{threshold_mode}{a character string indicating how the threshold is
determined and used. \code{cutoff} indicates that the threshold value is used as
a general maximum for all elements. \code{percentage} indicates that the cutoff
value is determined by the maximum element of each vector multiplied with the
threshold value. The default is set to \code{cutoff}.}

\item{expvar}{a character string indicating the method used for calculating
the explained variance. \code{approx} uses the explained variances of each
eigenvectors i.e. their eigenvalues. \code{exact} uses the variance of each variable.}

\item{check}{a character string indicating if only rows, or if rows as well as columns
are used to detect the underlying block structure. \code{rows} checks if the rows fullfill
the required structure. \code{rnc} checks if rows and columns fullfill the required structure.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
single or list of pla class containing the following attributes:
\item{x}{
  a numeric matrix or data frame which equals the input of \code{x}.
}
\item{c}{
  a numeric matrix or data frame which is the covariance or correlation
  matrix based on the input of \code{cov}.
}
\item{loadings}{
  a matrix of variable loadings (i.e. a matrix containing the
  eigenvectors of the dispersion matrix).
}
\item{threshold}{
  a numeric value which equals the input of \code{thresholds}.
}
\item{threshold_mode}{
  a character string which equals the input of \code{threshold_mode}.
}
\item{blocks}{
  a list of blocks which are identified by principal loading analysis.
}
See \url{https://www.sciencedirect.com/science/article/pii/S0047259X21000324} and
\url{https://dl.acm.org/doi/10.1145/3475827.3475832} for more information.
}
\description{
This function performs a principal loading analysis on the given data matrix.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9)
)
pla(data)

}
