% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{pla}
\alias{pla}
\title{Principle Loading Analysis

This function performs principle loading analysis on a data set.}
\usage{
pla(
  x,
  cor = FALSE,
  scaled_ev = FALSE,
  thresholds = 0.33,
  threshold_mode = "cutoff",
  expvar = "approx",
  check = "rnc",
  ...
)
}
\arguments{
\item{x}{a numeric matrix or data frame which provides the data for the
principal loading analysis.}

\item{cor}{a logical value indicating whether the calculation should use the
correlation or the covariance matrix. The default is set to "TRUE" indicating
the use of the correlation matrix.}

\item{scaled_ev}{a logical value indicating whether the eigenvectors should
be scaled. The default is set to "FALSE".}

\item{thresholds}{a numeric or list of numeric used to determine "small"
values inside the eigenvectors. If multiple values are given, a list of pla
results will be returned. The default is set to 0.33.}

\item{threshold_mode}{a character string indicating how the threshold is
determined and used. "cutoff" indicates that the threshold value is used as
a general maximum for all elements. "percentage" indicates that the cutoff
value is determined by the maximum element of each vector multiplied with the
threshold value. The default is set to "cutoff".}

\item{expvar}{a character string indicating the method used for calculating
the explained variance."approx" indicates the use of an approximation.
"exact" indicates the exact calculation. The default is set to "approx".}

\item{check}{a character string indicating if only rows or rows and columns
are used."rows" checks if the rows fullfill the required structure. "rnc"
checks if rows and columns fullfill the required structure. The default is
set to "rnc".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
single or list of pla class containing the following attributes:
\item{x}{
  a numeric matrix or data frame which equals the input of 'x'.
}
\item{c}{
  a numeric matrix or data frame which is the covariance or correlation
  matrix based on the input of 'cov'.
}
\item{loadings}{
  a matrix of variable loadings (i.e., a matrix whose columns contain the
  eigenvectors).
}
\item{threshold}{
  a numeric value which equals the input of thresholds at the corresponding
  position.
}
\item{threshold_mode}{
  a character string which equals the input of threshold_mode.
}
\item{blocks}{
  a list of blocks which are identified through the principle loading
  analysis.
}
See \url{https://arxiv.org/pdf/2007.05215.pdf},
\url{https://arxiv.org/pdf/2102.09912.pdf} for more information.
}
\description{
Principle Loading Analysis

This function performs principle loading analysis on a data set.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9)
)
pla(data)

}
