% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{print.pla}
\alias{print.pla}
\title{Print Function for pla S3}
\usage{
\method{print}{pla}(x, ...)
}
\arguments{
\item{x}{a pla object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A pla object which equals the input of \code{x}.
}
\description{
Prints the blocks, threshold, threshold_mode and the loadings.
}
\examples{
if(requireNamespace("AER")){
require(AER)
data("OECDGrowth")

pla_obj = pla(OECDGrowth,cor=TRUE,thresholds = 0.5)
print(pla_obj)
}

}
