% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{print.pla}
\alias{print.pla}
\title{Print Function for pla S3

Prints the blocks and threshold inside the object.}
\usage{
\method{print}{pla}(x, ...)
}
\arguments{
\item{x}{a pla object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Function for pla S3

Prints the blocks and threshold inside the object.
}
\examples{
data <- data.frame(
a = c(1:3),
b = c(4:6),
c = c(7:9)
)
obj <- pla(data)
print(obj)

}
