% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{spla}
\alias{spla}
\title{Sparse Principal Loading Analysis}
\usage{
spla(
  x,
  method = c("PMD", "SPCA"),
  para,
  cor = FALSE,
  criterion = "corrected",
  rho = 1e-06,
  max.iter = 200,
  trace = FALSE,
  eps.conv = 0.001,
  orthogonal = FALSE,
  check = "rnc",
  ...
)
}
\arguments{
\item{x}{a numeric matrix or data frame which provides the data for the
principal loading analysis.}

\item{method}{chooses the methods to calculate the sparse loadings.
\code{PMD} uses the method from REF(WittenEtAl2009) and \code{SPCA} uses the
method from REF(Zou2006).}

\item{para}{when \code{method = "PMD"}: an integer giving the bound for the
L1 regularization. When \code{method = "SPCA"}: a vector containing the
penalization parameter for each variable.}

\item{cor}{a logical value indicating whether the calculation should use the
correlation or the covariance matrix.}

\item{criterion}{a character string indicating if the weight-corrected
evaluation criterion (CEC) or the evaluation criterion (EC) is used.
\code{corrected} changes the loadings to weight all variables equally while
\code{normal} does not change the loadings and therefore weights the
variables differently.}

\item{rho}{penalty parameter. We refer to REF(Zou2006) and REF(Bauer22) for
a more elaborate explanation.}

\item{max.iter}{maximum number of iterations.}

\item{trace}{a logical value indicating if the progress is printed.}

\item{eps.conv}{a numerical value as convergence criterion.}

\item{orthogonal}{a logical value indicating if the sparse loadings are
orthogonalized.}

\item{check}{a character string indicating if only rows or rows as well as
columns are used to detect the underlying block structure. \code{rows} checks
if the rows fulfill the required structure. \code{rnc} checks if rows and
columns fulfill the required structure.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
single or list of \code{pla} class containing the following attributes:
\item{x}{
  a numeric matrix or data frame which equals the input of \code{x}.
}
\item{EC}{
  a numeric vector that contains the weight-corrected evaluation criterion
  (CEC) if \code{criterion="corrected"} and the evaluation criterion (EC) if
  \code{criterion="normal"}.
}
\item{loadings}{
  a matrix of variable loadings (i.e. a matrix containing the sparse
  loadings).
}
\item{blocks}{
  a list of blocks which are identified by sparse principal loading analysis.
}
\item{W}{
  a matrix of variable loadings used to calculate the evaluation criterion.
  If \code{criterion="corrected"}, \code{W} contains an orthogonal matrix
  with equal weights in the first column of each loading-block. If
  \code{criterion="normal"}, \code{W} equals \code{loadings}.
}
See REF(Bauer22) for more information.
}
\description{
This function performs the sparse principal loading analysis
provided by REF(Bauer22) on the given data matrix. The corresponding sparse
loadings are calculated either using \code{PMD} from the \code{PMA} package
or using \code{spca} from the \code{elasticnet} package.
}
\examples{
spla(USArrests, para=c(0.5,0.5,0.5,0.5), cor=TRUE)

## we obtain two blocks:
## 1x1 (Urbanpop) and 3x3 (Murder, Aussault, Rape).
## The large EC indicates that the given structure is reasonable.

spla(USArrests, para=c(0.5,0.5,0.7,0.5), cor=TRUE)

## we obtain three blocks:
## 1x1 (Urbanpop), 1x1 (Rape) and 2x2 (Murder, Aussault).
## The mid-ish EC for (Murder, Aussault) indicates that the found structure
## might not be adequate.

}
