% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla.R
\name{spla}
\alias{spla}
\title{Sparse Principal Loading Analysis}
\usage{
spla(
  x,
  type = "data",
  para = 0.5,
  cor = FALSE,
  orthogonal = TRUE,
  thresholds = 0.01,
  threshold_mode = "cutoff",
  check = "rnc",
  ...
)
}
\arguments{
\item{x}{a numeric matrix which can either be the data matrix or the covariance/
correlation matrix for sparse principal loading analysis.}

\item{type}{a character string. \code{data} indicates that \code{x} is the data matrix, and
\code{dispersion} indicates that \code{x} is the covariance or correlation matrix.}

\item{para}{a numeric value that gives the weight factor for sparsity.}

\item{cor}{a logical value indicating whether the calculation should use the
correlation or the covariance matrix. This option is only effective if \code{type = "data"}.}

\item{orthogonal}{a logical value indicating whether the sparse eigenvectors should be
orthogonal or not. If true, SPLA proceeds according to \insertRef{Benidis.2016}{prinvars} otherwise
according to \insertRef{Zou.2006}{prinvars}.}

\item{thresholds}{a numeric value or list of numeric values. All values of the
sparse eigenvectors that are less or equal to the threshold are set zo zero. If
multiple values are given, a list of SPLA results will be returned.}

\item{threshold_mode}{a character string indicating how the threshold is
determined and used. \code{cutoff} indicates the usage of a threshold value.
\code{percentage} indicates that the cutoff value is determined by the maximum
element of each vector multiplied with the threshold value. This option is only
effective if the parameter orthogonol is set to FALSE.}

\item{check}{a character string indicating if only rows or rows as well as columns
are used to detect the underlying block structure. \code{rows} checks if the rows fulfill
the required structure. \code{rnc} checks if rows and columns fulfill the required structure.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
single or list of \code{pla} class containing the following attributes:
\item{x}{
  a numeric matrix or data frame which equals the input of \code{x}.
}
\item{loadings}{
  a matrix of variable loadings (i.e. a matrix containing the
  eigenvectors of the dispersion matrix).
}
\item{threshold}{
  a numeric value which equals the input of \code{thresholds}.
}
\item{threshold_mode}{
  a character string which equals the input of \code{threshold_mode}.
}
\item{blocks}{
  a list of blocks which are identified by principal loading analysis.
}
}
\description{
Performs a sparse principal loading analysis (SPLA) on the given data matrix.
}
\examples{
if(requireNamespace("AER")){
require(AER)
}

}
