% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{str,Block-method}
\alias{str,Block-method}
\title{Block - str}
\usage{
\S4method{str}{Block}(object)
}
\arguments{
\item{object}{block.}
}
\value{
A string representing the Block.
}
\description{
Generic function to create a string out of the blocks structure.
}
\examples{
block <- new("Block", features=c(2, 5), explained_variance=0.03)
str(block)
}
